// src/types/todayMeetingType.ts
type TodayMeetingType = {
  groupId: number;
  memberId: number;
  nickname: string;
  image: string;
  title: string;
  name: string;
  food: string;
  date: string;
  time: string;
  maximum: number;
  current: number;
  storeName: string;
  storeAddress: string;
  createdDate: string;
};

export default TodayMeetingType;